first version of the MlosCoreOptimizer._to_df() tests
