+++ mlos_bench/mlos_bench/tests/optimizers/conftest.py
@@ -6,6 +6,8 @@
 Test fixtures for mlos_bench optimizers.
 """
 
+from typing import List
+
 import pytest
 
 from mlos_bench.tunables.tunable_groups import TunableGroups
@@ -13,6 +15,39 @@ from mlos_bench.optimizers.mock_optimizer import MockOptimizer
 from mlos_bench.optimizers.mlos_core_optimizer import MlosCoreOptimizer
 
 from mlos_bench.tests import SEED
+
+
+@pytest.fixture
+def mock_configs() -> List[dict]:
+    """
+    Mock configurations of earlier experiments.
+    """
+    return [
+        {
+            'vmSize': 'Standard_B4ms',
+            'idle': 'halt',
+            'kernel_sched_migration_cost_ns': 50000,
+            'kernel_sched_latency_ns': 1000000,
+        },
+        {
+            'vmSize': 'Standard_B4ms',
+            'idle': 'halt',
+            'kernel_sched_migration_cost_ns': 40000,
+            'kernel_sched_latency_ns': 2000000,
+        },
+        {
+            'vmSize': 'Standard_B4ms',
+            'idle': 'mwait',
+            'kernel_sched_migration_cost_ns': 100000,
+            'kernel_sched_latency_ns': 3000000,
+        },
+        {
+            'vmSize': 'Standard_B2s',
+            'idle': 'mwait',
+            'kernel_sched_migration_cost_ns': 200000,
+            'kernel_sched_latency_ns': 4000000,
+        }
+    ]
 
 
 @pytest.fixture
@@ -106,7 +141,6 @@ def flaml_opt_max(tunable_groups: TunableGroups) -> MlosCoreOptimizer:
 # normally calculated as a percentage of the max_iterations and number of
 # tunable dimensions, so for now we set the initial random samples equal to the
 # number of iterations and control them with a seed.
-
 
 SMAC_ITERATIONS = 10
